Add reusable connection building
This includes a new ReusableConnectionBuilder class, which requires
an instance of the new ExecutionContextFactory interface.  Ideally,
it should now be possible to easily create new connections from a
ReusableConnectionBuilder, with a properly written
ExecutionContextFactory, using only a socket.  This was implemented
as I expect it to be very useful for the upcoming jenjin-server
implementation.